Fixed tests and edge case, added Factor
StepRange is needed in some cases of x:dx:y if dx is not the same datatype
as x and y are. Added case where dx is set to 1::Int

Added factors datatype  the datatype that will do most of the magic.
Had trouble with parametrics. Should be worked out. Added bunch of helper
functions. Also, no tests yet